teensy3: usb_serial: Fix port 2/3 mismatch
The code for the second port was accidentally expanded into
usb_serial3.c, while the code for the third port was expanded in
usb_serial2.c.

Fixes: cd4c30fde5d09d0b ("Expand all USB serial 2 & 3 stuff (needed for Arduino IDE use)")
